+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Keep track of IP reputations
 * Correctly remove site instances from the `dispatcher.domains` object
 * Make sure `SiteDispatcher#getTestPort()` does not get stuck in an infinite loop
+* Make sure regexes do not accumulate after updating a site
 
 ## 0.5.3 (2023-10-19)